:building_construction: Adding failsafe for auto-shutdown

+++ maginkcal.py
@@ -149,11 +149,14 @@ def main():
         )
     )
     if isShutdownOnComplete:
-        logger.info("Shutting down safely.")
-        import os
-
-        os.system("sudo shutdown -h now")
-
+        # implementing a failsafe so that we don't shutdown when debugging
+        # checking if it's 6am in the morning, which is the time I've set PiSugar to wake and refresh the calendar
+        # if it is 6am, shutdown the RPi. if not 6am, assume I'm debugging the code, so do not shutdown
+        if currDatetime.hour == 6:
+            logger.info("Shutting down safely.")
+            import os
+            os.system("sudo shutdown -h now")
+            logger.info("Shutting down safely.")
 
 if __name__ == "__main__":
     main()
